+++ demo/components/universal/Mono/Mono.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import { Text } from 'componentry'
+
+const Mono = props => <Text as="span" monospace {...props} />
+
+export default Mono