+++ demo/components/universal/Mono/Mono.registry.js
@@ -1,0 +1,2 @@
+export const registryName = 'Mono'
+export { default } from './Mono'